Updated ImageType enum and relevant instances to include 'Mountain'
The 'Mountain' enum was added to the ImageType as a new type of image that the AI can recognize and denoise. Accordingly, 'Mountain' images were added in the AiDenoising project to be copied to the output directory. Instances where the ImageType was specified as 'House' was updated to 'Mountain'. This enables the application to work with a more diverse set of images and improves its flexibility.